Aiservice/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9         # file uploads

# ── Settings ──────────────────────────────────────────────────────────────────
pydantic-settings==2.2.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Security / Auth ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Vector Store ──────────────────────────────────────────────────────────────
qdrant-client==1.9.1

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[hiredis]==5.0.4

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers==2.7.0

# ── LLM (CPU path — Phi-2) ────────────────────────────────────────────────────
transformers==4.41.0
torch>=2.3.0
accelerate==0.30.0

# ── Document Parsing ──────────────────────────────────────────────────────────
pypdf2==3.0.1
python-docx==1.1.2

# ── Web Scraping ──────────────────────────────────────────────────────────────
httpx==0.27.0
beautifulsoup4==4.12.3

# ── Optional: GPU / GGUF path (uncomment if GPU available) ────────────────────
llama-cpp-python
# torch==2.3.0                  # full CUDA build
